requirements.txt
customtkinter
Pillow
tifffile
imagecodecs
